+++ src/Orchestra/Foundation/Application.php
@@ -1,6 +1,7 @@
 <?php namespace Orchestra\Foundation;
 
 use Exception;
+use Orchestra\Memory\Provider;
 
 class Application extends Abstractable\RouteManager
 {
@@ -72,7 +73,7 @@ class Application extends Abstractable\RouteManager
         }
 
         $this->services['orchestra.memory'] = $memory;
-        $app['events']->fire('orchestra.started');
+        $this->prepare($memory);
 
         return $this;
     }
@@ -88,9 +89,20 @@ class Application extends Abstractable\RouteManager
         $this->services['orchestra.menu'] = $this->app['orchestra.widget']->make('menu.orchestra');
         $this->services['app.menu']       = $this->app['orchestra.widget']->make('menu.app');
         $this->services['orchestra.acl']  = $this->app['orchestra.acl']->make('orchestra');
+    }
 
-        // Setup default drivers.
+    /**
+     * Prepare application.
+     *
+     * @param  \Orchestra\Memory\Provider   $memory
+     * @return void
+     */
+    protected function prepare(Provider $memory)
+    {
         $this->app['orchestra.notifier']->setDefaultDriver('orchestra');
+        $this->app['orchestra.mail']->attach($memory);
+
+        $this->app['events']->fire('orchestra.started', array($memory));
     }
 
     /**