+++ src/views/extensions/_table.blade.php
@@ -1,63 +1,63 @@
-<?php 
+<?php
 
 use Illuminate\Support\Fluent;
 use Orchestra\Support\Facades\Extension; ?>
 <table class="table table-striped">
 	<thead>
 		<tr>
-			<th><?php echo trans('orchestra/foundation::label.extensions.name'); ?></th>
-			<th><?php echo trans('orchestra/foundation::label.description'); ?></th>
+			<th>{{ trans('orchestra/foundation::label.extensions.name') }}</th>
+			<th>{{ trans('orchestra/foundation::label.description') }}</th>
 		</tr>
 	</thead>
 	<tbody>
-		<?php if (empty($extensions)) : ?>
+		@if (empty($extensions))
 		<tr>
-			<td colspan="2"><?php echo trans('orchestra/foundation::label.no-extension'); ?></td>
+			<td colspan="2">{{ trans('orchestra/foundation::label.no-extension') }}</td>
 		</tr>
-		<?php else : 
-		foreach ($extensions as $name => $extension) :
-			$extension = new Fluent($extension); ?>
+		@else
+		@foreach ($extensions as $name => $extension)
+			<? $extension = new Fluent($extension); ?>
 		<tr>
 			<td>
 				<strong>
-					<?php 
+					<?
 					$active  = Extension::activated($name);
 					$started = Extension::started($name);
-					$uid     = str_replace('/', '.', $name);
+					$uid     = str_replace('/', '.', $name); ?>
 
-					if ( ! ($started)) :
-						echo $extension->name;
-					else : ?>
-						<a href="<?php echo handles("orchestra::extensions/configure/{$uid}"); ?>">
-							<?php echo $extension->name; ?>
+					@if (! ($started))
+						{{ $extension->name }}
+					@else
+						<a href="{{ handles("orchestra::extensions/configure/{$uid}") }}">
+							{{ $extension->name }}
 						</a>
-					<?php endif; ?>
+					@endif
 				</strong>
 				<div class="pull-right btn-group">
-					<?php if ( ! ($started or $active)) : ?>
-						<a href="<?php echo handles("orchestra::extensions/activate/{$uid}"); ?>" class="btn btn-primary btn-mini">
-							<?php echo trans('orchestra/foundation::label.extensions.actions.activate'); ?>
+					@if (! ($started || $active))
+						<a href="{{ handles("orchestra::extensions/activate/{$uid}") }}" class="btn btn-primary btn-mini">
+							{{ trans('orchestra/foundation::label.extensions.actions.activate') }}
 						</a>
-					<?php else : ?>
-						<a href="<?php echo handles("orchestra::extensions/deactivate/{$uid}"); ?>" class="btn btn-warning btn-mini">
-							<?php echo trans('orchestra/foundation::label.extensions.actions.deactivate'); ?>
+					@else
+						<a href="{{ handles("orchestra::extensions/deactivate/{$uid}") }}" class="btn btn-warning btn-mini">
+							{{ trans('orchestra/foundation::label.extensions.actions.deactivate') }}
 						</a>
-					<?php endif; ?>
+					@endif
 
 				</div>
 			</td>
 			<td>
 				<p>
-					<?php echo $extension->description; ?>
+					{{ $extension->description }}
 				</p>
 
 				<span class="meta">
-					<?php echo trans('orchestra/foundation::label.extensions.version', array('version' => $extension->version )); ?> |
-					<?php echo trans('orchestra/foundation::label.extensions.author', array('author' => sprintf('<a href="%s" target="_blank">%s</a>', $extension->url ?: '#', $extension->author))); ?>
+					{{ trans('orchestra/foundation::label.extensions.version', array('version' => $extension->version )) }} |
+					{{ trans('orchestra/foundation::label.extensions.author', array('author' => sprintf('<a href="%s" target="_blank">%s</a>', $extension->url ?: '#', $extension->author))) }}
 				</span>
 			</td>
 		</tr>
-		<?php endforeach;
-		endif; ?>
+		@endforeach
+		@endif
 	</tbody>
 </table>